move player engine into a separate runnable to make it more flexible in the way we use it
